Fast af boii

+++ TestStandSoftware/LoadCell.h
@@ -25,6 +25,6 @@ void initLoad(void);
  * IN:            Load cell number to read
  * OUT:           float value with the load cell measurement of the given sensor number
  */
-float readLoad();
+int readLoad();
 
 #endif
